Split ColorFade out, refactored the title elements
