+++ src/screens/Base.py
@@ -6,6 +6,11 @@ MIT license, see LICENSE.md for details.
 
 import pygame
 
+from ..ui import ImageButton
+from ..ui import Label
+
+WHITE = pygame.colordict.THECOLORS['white']
+
 
 class Base:
     def __init__(self, game):
@@ -17,6 +22,9 @@ class Base:
 
         self.can_exit = False  # Time for this screen to exit?
         self.next_screen = None  # Next screen to display.
+
+        self.has_title = False
+        self.title_ui = []  # For Title elements.
 
     def draw(self):
         ''' Draw the screen's contents.
@@ -43,46 +51,22 @@ class Base:
     def mouseup(self, event):
         pass
 
+    def addTitle(self):
+        title_text = self.game.text.getText('title')
+        skelly_text = self.game.text.getText('skelly_title')
+        subtitle_text = title_text['subtitle_text']
 
-class ColorFade:
-    def __init__(self, rgba1, rgba2, duration):
-        self.rgba = tuple(rgba1)
-        self.start_rgba = rgba1
-        self.end_rgba = rgba2
-        self.duration = duration
+        self.title_ui = [
+            ImageButton(0, 0, self.game.resources['images']['skelly_title']),
+            Label(self.game.screen_width / 2, 40, skelly_text, self.game.resources['fonts']['skelly_title'],
+                  WHITE, 'centre'),
+            Label(self.game.screen_width / 2, 220, subtitle_text, self.game.resources['fonts']['default_mono'],
+                  WHITE, 'centre'),
+        ]
 
-        self.dt = 0
-        self.done = False
+        self.has_title = True
 
-        info = pygame.display.Info()
-        self.screen_rect = pygame.Rect(0, 0, info.current_w, info.current_h)
-        self.texture = pygame.Surface((self.screen_rect.width, self.screen_rect.height))
-        self.texture.fill(self.rgba)
-        self.texture.set_alpha(self.rgba[-1])
-
-    def getColor(self):
-        ''' Return the current colour.
-        '''
-        return self.rgba
-
-    def isDone(self):
-        return self.done
-
-    def update(self, dt):
-        if self.done:
-            return
-
-        self.dt = self.dt + dt
-        if self.dt > self.duration:
-            self.rgba = tuple(self.end_rgba)
-
-            self.done = True
-            return
-
-        diff = self.dt / self.duration
-        self.rgba = tuple(map(lambda a, b: a + (b - a) * diff, self.start_rgba, self.end_rgba))
-        self.texture.fill(self.rgba)
-        self.texture.set_alpha(self.rgba[-1])
-
-    def draw(self):
-        pygame.display.get_surface().blit(self.texture, self.screen_rect)
+    def drawTitle(self):
+        if self.has_title:
+            for item in self.title_ui:
+                item.draw()
