+++ src/screens/__init__.py
@@ -4,14 +4,13 @@ By Chris Herborth (https://github.com/Taffer)
 MIT license, see LICENSE.md for details.
 '''
 
-from .Base import Base, ColorFade
+from .Base import Base
 from .Journey import Journey
 from .Presents import Presents
 from .Title import Title
 
 __all__ = [
     'Base',
-    'ColorFade',
     'Journey',
     'Presents',
     'Title',
