+++ src/ui/__init__.py
@@ -6,12 +6,14 @@ MIT license, see LICENSE.md for details.
 
 from .Base import Base
 from .Button import Button
+from .ColorFade import ColorFade
 from .ImageButton import ImageButton
 from .Label import Label
 
 __all__ = [
     'Base',
     'Button',
+    'ColorFade',
     'ImageButton',
     'Label',
 ]
